requirements.txt
numpy
pandas
pyyaml
streamlit
streamlit-authenticator
